+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Unreleased]
 
-Nothing noticeable unreleased.
+- Indexer: by default, sleep for 1 second between requests during crawl.
 
 ## v2025.8